[#2396] Remove auth0.js and move logic into auth.js [client]
